+++ src/components/feature/ProductItem/CartButton/index.tsx
@@ -1,6 +1,6 @@
 import { ProductItem } from 'types/productItem';
 
-import styles from './CartButton.module.scss';
+import styles from './index.module.scss';
 
 type CartButtonProps = {
   id: ProductItem['id'];
