+++ src/components/feature/ProductItem/Price/index.module.scss
