+++ src/components/feature/ProductItem/Thumbnail/index.module.scss
